requirements.txt
requests
gunicorn
flask
logzero
PyYAML
tqdm
beautifulsoup4
